pyproject.toml
[project]
name = "pier2"
version = "0.1.0"
description = "Because all you touch and all you see is all your life will ever be."
authors = [
    {name = "Pink Floyd"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyaml (>=25.1.0,<26.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
requests = "^2.32.3"
httpx = "^0.28.1"
sqlmodel = "^0.0.22"
numpy = "^2.2.2"
pandas = "^2.2.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

